refactor: Layers, Components 디렉토리 구조화 [#98]
- Kanban > Layers 이동
